+++ app/Views/admin/posts/edit.php
@@ -242,7 +242,12 @@
                 title,
                 content
             })
-        }).then(r => r.json()).then(data => data.forEach(addTag)).finally(() => {
+        }).then(r => r.json()).then(data => {
+            // Reset tag lama sebelum menambahkan hasil saran
+            tagContainer.querySelectorAll('.tag-badge').forEach(b => b.remove());
+            updateTagsInput();
+            data.forEach(addTag);
+        }).finally(() => {
             this.disabled = false;
             this.innerHTML = '<i class="fa-solid fa-fw fa-wand-magic-sparkles mr-2"></i>Sarankan Tag';
         });